fix error when msg from callback is null

+++ lib/QueueClient.ts
@@ -120,7 +120,7 @@ export class QueueClient {
 
         // listen for callback
         return await channel.consume(queue.queue, async (msg) => {
-            if (msg.properties.correlationId == correlationID) {
+            if (msg && 'properties' in msg && msg.properties.correlationId == correlationID) {
                 const msgJson = JSON.parse(msg.content.toString());
                 await channel.deleteQueue(queue.queue);
                 await channel.close();
